Only fetch requested ops for `getOps`
# Introduction

At the moment, when calling `getOps`, `sharedb-mongo` actually fetches
all the ops from `from` to the current version.

That means that if I have a document with 1,000 versions, and I only
ask for ops 0-10, we still fetch all 1,000 ops. This is clearly
inefficient, and on documents with large numbers of ops, this fetch can
take a long time. Anecdotally, it takes ~2s to fetch ops 0-10 of
100,000 ops.

This change has an anecdotal performance increase of ~100x, now taking
~0.01s to fetch ops 0-10 of 100,000 ops.

# Background

We fetch all the ops from the current version, because we try to fetch a
"valid" chain of operations. Consider this case:

  1. I have a doc v1
  2. I submit an op v2
  3. I concurrently submit another op v2
  4. Both concurrent ops are committed to the database before the
     snapshot
  5. One of the commits successfully writes a snapshot
  6. The other snapshot is rejected, but its op is still committed
  7. I now try to get all ops. A naive `find` operation would return two
     ops for v2.

In theory, this case should actually already be [cleaned up][1], so I
can only assume that we are:

  - guarding against cases where the op cleanup fails
  - guarding against unknown inconsistencies in ops
  - guarding against stupid consumers mucking about in the db themselves
  - being very defensive

In order to therefore fetch a valid set of ops, we currently therefore
fetch the current snapshot, which has a reference stored to the op that
created it. By looking up that op, we can then check the op that
preceded _that_, and so on, forming a valid chain of ops that point to
one another.

The obvious downside of this is that we need to start at the current
snapshot and work backwards, necessitating a fetch of all the ops up to
the current version.

# Changes

This change attempts to allow us to only fetch the bare minimum number
of ops, whilst still maintaining integrity and discarding erroneous ops.

This is achieved by attempting to find the first op after our `to` op
that has a unique version. The assumption here is that any op with a
unique version is inherently valid, because it had no collisions.

For example, consider the case where I have a document with some ops:

  - v1: unique
  - v2: unique
  - v3: collision 3
  - v3: collision 3
  - v4: collision 4
  - v4: collision 4
  - v5: unique
  - v6: unique
  - ...
  - v1000: unique

If I want to fetch ops v1-v3, then we:

  - look up v4
  - find that v4 is not unique
  - look up v5
  - see that v5 is unique and therefore assumed valid
  - look backwards from v5 for a chain of valid ops

This way we don't need to fetch all the ops from v5 to the current
version.

In the case where a valid op cannot be determined, we still fall back to
fetching all ops and working backwards from the current version.

# Further work

Note that this change *only* affects the `getOps` method, notably not
touching either `getOpsToSnapshot` or `getOpsBulk`.

[1]: https://github.com/share/sharedb-mongo/blob/2d579ddb80781e987707076e932cd4e01ca066ef/index.js#L189-L193