✅ Fix `npm test`
This is broken in a fresh install:

```
npm install
npm test
```

This is because `npm` will fetch `mongodb@5` (the latest it can), but
running `npm test` without the `_SHAREDB_MONGODB_DRIVER` variable set
means that the map reduce test still runs, when `mongodb@5` doesn't
support it.

This change tweaks the test to check if the function is defined, instead
of relying on the environment variable.